Enabling dep to pickup Uber internal phab repos.
Summary:
Added a deducer to be invoked everytime there is an import path that starts with code.uber.internal. The deducer interface requires deduceRoot and deduceSource functions that return what the root repo is for the import path and also where the package source is.

Also added a workaround for a fix for dep not reading glide.yaml correctly for branches. "origin/master" does not match the "master" branch in phab. Added test cases to test the scenarios.

These changes enable us to run dep init on catalyst-backend repo.

Differential Revision: https://code.uberinternal.com/D1139795