Mechanism to enable import of custom configuration for dep migration
Summary:
Glide dependency resolution is not optimal and so dep is not
guaranteed to resolve just because glide does. This change will
make it easier to transition from glide to dep. It allows a file
called customConfig.yaml that can be used to set overrides and
manual conflict resolution.

Reviewers: monty, momen, vit

Reviewed By: momen

Tags: #build_platform, #build_platform_sprint_10_16_-_10_20

Maniphest Tasks: T1274059

Differential Revision: https://code.uberinternal.com/D1286543