Adding the golang gitolite mirroring logic from glide
Summary: Adapted the gitolite mirroring logic from glide for golang.org sources. Dep doesn't have a dedicated deducer for golang.org, so the changes were made in the generic deducer, that way if anything fails with the gitolite logic, it will fall back onto what was previously there with dep.

Reviewers: vit, momen, chinmayr

Reviewed By: chinmayr

Differential Revision: https://code.uberinternal.com/D1316907