Add deducer for googlesource and optimize mirroring logic
Summary:
The golang/x imports save source code inside googlesource
so the googlesource regex redirects to the golang deducer.
The mirroring logic performs ls-remote for each repo
which makes it really slow. The repo only needs to be
mirrored once so we will default to the repo existing
internally. Only mirror the repo async in the background.

Reviewers: prabhuk, monty, momen, vit

Reviewed By: vit

Differential Revision: https://code.uberinternal.com/D1376499