Add bootstraping overrides feature and fix bug
Summary:
We have realized the need to boostrap a basic
overrides file that lists the common overrides
needed for a bunch of repos. This commit adds
the ability to do that, and also adds the
golang.org/x/net repo source override that
is identified to fail a lot of repos.

Reviewers: momen, monty, vit

Reviewed By: vit

Tags: #build_platform, #build_platform_sprint_12_25_-_01_05

Differential Revision: https://code.uberinternal.com/D1423111