Add ability to read and bootstrap basic exclude dirs
Summary:
There are some Uber repos that specify exclude dirs
such as .gen, .tmp and _templates in their glide
manifest but majority of the repos do not. This
causes issues which is why these excludes will be
part of the bootstrapped dep config. The ability to
honor these exclude dirs from the dep config is also
part of this commit.

Reviewers: vit, monty, momen

Reviewed By: monty

Tags: #build_platform, #build_platform_sprint_1_8_-_1_19

Differential Revision: https://code.uberinternal.com/D1429169