Handle gitolite repos that don't end in .git and improve panic
Summary:
There is a panic occuring as part of the solver when it runs
out of versions for a dependency. The underlying issue is that
some gitolite repos have ambigious repo roots, so we fail to
mirror the repo correctly. Since there is no repo mirrored,
the solver is unable to get a list of versions which causes
it to panic. This commit includes a fix for deducing ambigious
gitolite roots and also panic if the gitolite repo fails to
mirror correctly.

Reviewers: vit, monty, momen, zplin

Reviewed By: vit

Tags: #build_platform, #build_platform_sprint_1_8_-_1_19

Differential Revision: https://code.uberinternal.com/D1431955