Adding error reporting for metrics
Summary: This adds reporting for errors in solve_failures.go. It adds the reporting to the Error() functions for each error, which I'm not entirely sure if that will ensure exactly one call per error occurrence. I'll test this out and if it doesn't work, I'll add it to a better place.

Reviewers: vit, momen, chinmayr

Reviewed By: chinmayr

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1436899