Prevent multiple errors being reported and only create a single metrics reporter per run
Summary:
We are creating a new metrics reporter everytime we are reporting a timer or a error
counter metric. This was incorrect as we should only be creating one per run and post
all metrics using that reporter to multiple scopes. Also made the change to only post
one error per run instead of nested errors.

Reviewers: vit, osama, monty, momen

Reviewed By: vit, monty, momen

Subscribers: momen

Tags: #build_platform

Differential Revision: https://code.uberinternal.com/D1471535