Avoid hitting github before gitolite for go.uber.org repos to prevent conflict
Summary:
go.uber.org internal repos lead to an external path of github, and the same external github
repos lead to an external remote path of github (but an internal path of gitolite repo).
When we get these overlapping github paths, we end up setting the same remote URL for both.
The same remote path for both cases is invalid and so we need to ignore the external github
path in the case when it can also be defined as a go.uber.org path.

Reviewers: vit, osama

Reviewed By: osama

Tags: #build_platform

Differential Revision: https://code.uberinternal.com/D1568919