Remove special case for go.uber.org and fix the issue with duplicate keys mapping to same source gateway.
Summary:
go.uber.org maps to the same remote that those gitolite repos map to. We special cased gitolite to not
hit github first in those cases. The underlying problem is that the source manager keys off the remote
URL. This change adds a field to the remote URL that can be kept the same even though the source is
pulled from the same remote, so there is no conflict in the hashmap of URL to source gateway.

Reviewers: osama, vit

Reviewed By: osama

Differential Revision: https://code.uberinternal.com/D1571045