Change the location of the dep cache to $HOME/.dep-cache
Summary:
This needs to be done so that go-build can reference the dep cache from its temporary gopath it creates (the .tmp folder). Having it in the home directory allows it to be used in different scenarios without being locked to a particular gopath the way it is now.

Tested manually and ran unit tests.

Reviewers: vit, osama

Reviewed By: vit

Differential Revision: https://code.uberinternal.com/D1580923