Keep glide manifest in sync with dep manifest for backwards compatibility
Summary: This change is needed to allow safe migration of libraries from glide to dep. This allows Dep to create and maintain a glide manifest, and keep it in sync with the dep manifest. This allows for backwards compatibility for your defenders to be able to read this glide manifest.

Reviewers: vit, osama, #build_platform

Reviewed By: vit, #build_platform

Differential Revision: https://code.uberinternal.com/D1597253