Fix curropt cache issues.
Summary:
1) Cloning/fetching killed midway can leave behind a partially cloned
cache folder. This required cleaning the cache and failures. This
commit fixes this issue by clearing the folder and recloning if it
is missing .git or .git/index.
2) The definition of existsLocally now has an added condition of
not only having a .git folder but also the .git/index present. If
not then the package is curropt and requires recloning.

Reviewers: vit

Reviewed By: vit

Tags: #build_platform

Differential Revision: https://code.uberinternal.com/D1628185