adding encoded relationships between nodes
Summary: The code in the CR is creating relationships between already-encoded nodes. It takes the output of a previous function call, which encodes the nodes through hashing, and then draws relationships between the hashed values by iterating over the given dependency tree. The output of mkRelationships will be used later to write to a file which will be output to the user to copy/paste into graphviz, much in the same way as the dep status -dot

Reviewers: chinmayr, zplin, vit, #build_platform

Reviewed By: chinmayr, zplin, #build_platform

Maniphest Tasks: T1790339

Differential Revision: https://code.uberinternal.com/D1768925