Improve logging around ignored constraints, fix metrics panic
Summary:
Dep only honors constraints from your direct/transient dependencies that are of their direct dependencies. Dep strongly enforces the rule that all repos should only be able to set a constraint on their direct dependency. Overrides are used for transient dependencies, and constraints are used for direct dependencies. If a repo sets a constraint for their transient or code-gen dependency (thriftrw) then Dep will read them in but ignore them as part of the solver process, causing you to believe that Dep is not honoring all constraints. This is by design to maintain some level of sanity.

Improving logging around this.

Reviewers: vit, zplin, cwest1

Reviewed By: vit, zplin

Tags: #build_platform

Differential Revision: https://code.uberinternal.com/D1779261