Enable ability to move lock input digest to a local hidden file
Summary:
This enables the ability to move the lock input digest to a local hidden file. The .gopkg.digest will only be local and not be pushed upstream. It will provide the same functionality the digest provides currently. This will be useful in the monorepo in avoiding lock conflicts upstream when only the digest has changed.

For existing repos: If there is a input digest present in the lock file, then everything keeps working as before.

Migration plan for monorepo: The way to remove the input digest from the lock is to create a .gopkg.digest file with the lock (add it to .gitignore) and move the digest from the lock to this new file. When devs in monorepo pull the lock without digest for the first time, the code will automatically create a .gopkg.digest file for them. The dep run will also start writing digest to the file since it now exists instead of adding it back to the lock.

Test Plan: Tested multiple usecases and migration strategy against microrepo (sre/udestroy) and monorepo. Verified there is no regression before and after migration in monorepo.

Reviewers: vit, zplin, zinzun, #build_platform

Reviewed By: vit, #build_platform

Maniphest Tasks: T1986805

Differential Revision: https://code.uberinternal.com/D2231771