Expose RawProject for ordered constraints/overrides (slice vs map), inplace of using Manifest
Summary: The dep.Manifest struct is derived from the RawManiest but converts slice to a map in the process, which makes reading and writing back the manifest out of order. The RawManifest needs to be exposed as the struct to read the manifest file into, to maintain order of constraints and overrides.

Reviewers: vit, zplin, zinzun, #build_platform

Reviewed By: zinzun, #build_platform

Maniphest Tasks: T2203815

Differential Revision: https://code.uberinternal.com/D2239705